requirements.txt

torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.3.0
matplotlib>=3.4.3
PyYAML>=5.4.1
wandb>=0.12.0
opencv-python>=4.5.3
tqdm>=4.62.3
scikit-learn>=0.24.2